fixed booking managements and client booking and supplierschedule and improve performance

+++ src/app/api/uploadthing/profile/change/client/route.ts
@@ -26,8 +26,12 @@ export const POST = async (req: Request) => {
 
             if (client.profile_key) {
 
-                const deleteProfile = await utapi.deleteFiles(client.profile_key)
-                if (!deleteProfile) return badRequestRes()
+                const clientProfile = await utapi.getFileUrls(client.profile_key)
+                
+                if (clientProfile) {
+                    const deleteProfile = await utapi.deleteFiles(client.profile_key)
+                    if (!deleteProfile) return badRequestRes()
+                }
 
                 const updateClientProfile = await prisma.client.update({
                     where: { id: client.id }, data: { profile_key: profile.key, profile_url: profile.url }
